Split test definition
